requirements/main.txt
Flask==2.3.2
openai==0.27.5
python-dotenv==1.0.0
yt-dlp==2023.3.4